requirements.txt
pandas
pandas_datareader
dash
plotly
flask
gunicorn
dash-bootstrap-components 
python-telegram-bot==20.7
python-dotenv
asyncio
datetime
